Slim V8DOMConfiguration::MethodConfiguration by refactoring per-world bindings.
This represents the main world version as a separate entry in the configuration list.
While this is larger for methods with PerWorldBindings, it's smaller for those without,
which are the vast majority.

Effect is a change of 10-15 KiB in the .data.rel.ro section on Linux x64.
Android Monochrome (32-bit ARM) APK size shrinks by 8 KiB.

Review-Url: https://codereview.chromium.org/2717183002
Cr-Original-Commit-Position: refs/heads/master@{#457922}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f43563b2d986dd686d4f20a14fc15fe02884489