Remove header ordering checks from check-webkit-style
This change removes checks for header ordering in check-webkit-style.
We use clang-format to format and lint our headers so it is unnecessary
for check-webkit-style to check this.

See AnimationSimTest.cpp in https://codereview.chromium.org/2757543002 for
something linted by clang-format that would fail check-webkit-style
if this feature weren't removed.

BUG=701250

Review-Url: https://codereview.chromium.org/2749513008
Cr-Original-Commit-Position: refs/heads/master@{#458009}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2a9fb0166f08fecc7b3b68deff1b6523f1610d23