MouseEventManager: reset drag state upon detaching.
If an ongoing drag detaches the frame of the dragged element, have
the MouseEventManager reset the drag state while handling the detachment
from the frame. Otherwise the DragState will retain a reference to
the drag source (and its document) for too long, triggering a reported
leak on shutdown.

R=
BUG=703057

Review-Url: https://codereview.chromium.org/2762613002
Cr-Original-Commit-Position: refs/heads/master@{#458116}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f6108b39ff0fa823e85bada2e56dbf89d4f9acad