DevTools: Fix Network panel warning icon display when offline/throttled
Fixes an issue whereby the warning icon on the tab for the Network panel would not appear when offline or the network was otherwised throttled.

This issue was a regression in https://codereview.chromium.org/2692653003.

The fix works in the Network panel and Performance panel when offline/throttling is changed.

The following issue http://crbug.com/702070 prevents the warning icon from being displayed when emulating offline from the Service Workers pane within the Applications panel. https://codereview.chromium.org/2757093002 should fix this.

BUG=700797

Review-Url: https://codereview.chromium.org/2758183002
Cr-Original-Commit-Position: refs/heads/master@{#458260}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 33bc6b04dccef0bc2c1066dad202474e35e6dda7