Fix oninstall-script tests for service worker
According to the spec and the WPT test, an exception occurring during oninstall
should not cause installation to fail. This change makes the test pass.

Note that this checks removal may affect the behavior of all ExtendableEvents;
however it is expected to be fine as explained in the bug.

BUG=685531

Review-Url: https://codereview.chromium.org/2734783002
Cr-Original-Commit-Position: refs/heads/master@{#458339}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a626eb843a373043835437074f4c7de51baf00f4