[css-align] Import the CSS Box Alignment tests from the CSSWG Test Suite
This patch imports the tests from the CSSWG Test Suite that verify the
correct parsing and resolved values of the place-content, place-items
and place-self shorthand properties.

Currently the all fail; the ones verifying the place-content shorthand
because of the new baseline-position syntax (issue #703584). The
rest fail because we still have not implemented the support for the
place-items and place-self shorthands (issue #668639)

BUG=703592

Review-Url: https://codereview.chromium.org/2762043004
Cr-Original-Commit-Position: refs/heads/master@{#458386}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7b60b0148294aa25648798b20bb527097c800baf