Refactor ImageFrame::setSizeAndColorSpace()
The function ImageFrame::setSizeAndColorSpace does multiple things.
It sets the size and color space & zero fills the data.

We want to be able to not zero fill the entire image. The image decoder
can fill the entire image frame with valid pixel data. And in the case
of partial images, the decoder knows which pixels do not contain image
data, allowing them to zero fill a smaller area.

But for now, just separate the concept of setSizeAndColorSpace() and
zeroFillPixelData().

No change in behavior, no new tests.

R=scroggo@chromium.org
BUG=701143

Review-Url: https://codereview.chromium.org/2749703002
Cr-Original-Commit-Position: refs/heads/master@{#458437}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b6b96c3de7c6db78762ce660b274514bd18a1c0b