Clean up DocumentWriter creation's FrameLoader interaction
There are multiple methods in FrameLoader that take part in the commit
and Document creation sequence, with less-than-clear purposes. Also,
these methods make little-to-no use of FrameLoader's members, instead
primarily interacting with Document, LocalFrameClient, and the
DocumentLoader that is driving the commit. Move a bunch of this logic
to DocumentLoader, since it's all tied to the loading of a specific
Document anyway.

BUG=

Review-Url: https://codereview.chromium.org/2751833002
Cr-Original-Commit-Position: refs/heads/master@{#458549}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ace017f26fd8cb8d123238e41c2f0e11f5daacd6