Revert of Add use count for click retargeting due to capture
Reason for revert:
Reverting all the changes related to the click target in capturing as they are splitted across multiple commits.

This reverts commit 18a3165ddb7326845a6028b1e3f93f7d347bde95.

Original issue's description
Since pointer capture changes the target of the
following events for that pointer it potentially
changes the target of the click. Considering the
changed target for calculating the click target
may cause some websites to break. So this CL
adds a usecount for that case. Note this is an
upper bound on the breakage as lots of pages
may not have any listener for these click events
or just not break due to this retargetting.

BUG=699933

TBR=dtapuska@chromium.org,mustaq@chromium.org,bokan@chromium.org

Review-Url: https://codereview.chromium.org/2769433002
Cr-Original-Commit-Position: refs/heads/master@{#458550}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3d84b135e186d7d22d9597aa752ddc8f345be102