Refactoring: Replace test_prefix_list variables with TestBaselineSet objects.
Purpose: I think that this improves the readability of the code, since
it replaces dicts with a specific format that are being passed around
with instances of a class, and that class has documentation and tests.

Also, this should make it easier to fix http://crbug.com/673966; I'm
thinking that TestBaselineSet could be expanded to specify port names
separately from build names, so that rebaseline_cl could specify for
example that it wants to get baselines for some test from some Build,
but that the baseline location should correspond to another port's
baseline location.

This is a refactoring CL, so behavior should not change. In this CL:
 - Adds a class called TestBaselineSet (I'm open to name suggestions...)
   which represents some combinations of tests and platforms to rebaseline.
 - Adds a test for that class.
 - Updates other tests.
 - Removes _log_test_prefix_list; this is replaced with __str__ on the
   new class.
 - Refactors iteration through the collection of tests/builds to rebaseline.
 - Removes redundant method test_execute_with_issue_number_from_branch
   in rebaseline_cl_unittest.

BUG=702819

Review-Url: https://codereview.chromium.org/2765863003
Cr-Original-Commit-Position: refs/heads/master@{#458592}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 09067d4bab0caeb8e1409148c9d6beabf762fd39