Web Animations: support cssOffset
When animating the CSS shorthand property 'offset', cssOffset is used
in Web Animations keyframes.

https://w3c.github.io/web-animations/#property-name-conversion

Note that when using CSS Animations or Web Animations on a shorthand property, and discrete animation applies to one of the longhand properties, that longhand property updates at 50% progress. When using CSS Transitions on a shorthand property, and discrete animation applies to one of the longhand properties, that longhand property updates immediately.

BUG=700534

Review-Url: https://codereview.chromium.org/2747523002
Cr-Original-Commit-Position: refs/heads/master@{#458651}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f543cf03508ae5cc38f69a1871b99308859a2d8