Make all setNeedsReattachLayoutTree happen from updateStyle.
The setNeedsReattachLayoutTree calls which were done outside of
updateStyle were all accompanied by a NeedsReattachStyleChange for the
node marking the ancestor chain for recalc. Therefore we can simply
mark for re-attachment inside the recalcStyle pass instead.

Enforce the invariant by introducing a DCHECK that the we are inside
updateStyleAndLayoutTree, and that the shadow DOM distribution is
up-to-date.

R=nainar@chromium.org,esprehn@chromium.org
BUG=595137

Review-Url: https://codereview.chromium.org/2760233004
Cr-Original-Commit-Position: refs/heads/master@{#458683}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d6f8a937c0d8fbfcbdfe7edc93255e454c5de3cf