Remove the webkit-patch "rebaseline-json" command.
Reason: The docstring for this command says that it was intended to
be used by other commands, but in reality it's not being used for any
other commands.

The unit tests were actually not specific to rebaseline-json, so they
have been kept with minor modifications.

Review-Url: https://codereview.chromium.org/2758293003
Cr-Original-Commit-Position: refs/heads/master@{#458790}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ab0b8e3ab6e7c70b740be07e63aaebdc0418a95c