Centralize more querySelector logic behind QuerySelectorCache.
We make the types to the functions more clear, move the empty selector
check inside, and handle the NthIndexCache automatically instead of
expecting the callers to do the setup.

This also unifies the exception messages for matches(), closest(),
querySelector(), and querySelectorAll() when an empty string is passed.

BUG=703900

Review-Url: https://codereview.chromium.org/2769783002
Cr-Original-Commit-Position: refs/heads/master@{#458942}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 146a91c44120872cb59aba6078b7cb40e4e8f4f6