Revalidate SMIL animation value after target change
When a target moves in the tree, the animation value was invalidated,
but never revalidated again when the element was reinserted.
Make sure to signal a revalidation if there's an active interval and the
the animation is running after the target element was reinserted.

BUG=701128

Review-Url: https://codereview.chromium.org/2763283002
Cr-Original-Commit-Position: refs/heads/master@{#459150}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f70d58032706670f094ed7ae73c2bb13259fc0b6