Image Capture: prune Photo{Capabilities/Settings} and add MediaTrackConstraintSet.pointsOfInterest
- This CL removes from PhotoCapabilities.idl the fields that are
not there anymore according to the spec [1], leaving just four
of them.  The corresponding code and LayoutTests are updated.
Same for PhotoSettings.idl [2].

- This CL also adds ConstrainPoint2D data type [3] and uses it
for MediaTrackConstraintSet.pointsOfInterest, wiring it down
to the mojo interface and in the mock.

- This is the one and only important change visible to the users.
(There is an Origin Trial going on for Image Capture in
M56-57-58, but these changes will land in 59.x.y.z, so shouldn't
affect them, in any case an update to blink-dev@ is being
brewed and there's a migration guide:
  https://tinyurl.com/image-capture-migration-guide
)

- No more MediaStreamTrack idl changes are expected after
this CL \o/

TEST= Layouttests etc updated.
BUG=700607

[1] https://w3c.github.io/mediacapture-image/##photocapabilities-section
[2] https://w3c.github.io/mediacapture-image/##photosettings-section
[3] https://w3c.github.io/mediacapture-image/#additional-constrainable-props

Review-Url: https://codereview.chromium.org/2773593004
Cr-Original-Commit-Position: refs/heads/master@{#459367}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f9e2de323e9265aee33ba9683ba98b09fec6769