[LayoutNG] Add NGLineBoxFragment
This patch adds NGLineBoxFragment, along with:
- NGPhysicalLineBoxFragment
- NGLineBoxFragmentBuilder
- NGTextFragmentBuilder for consistency

Part of code in NGLineBuilder was extracted to new classes, in
preparation of renaming it to NGLineLayoutAlgorithm.

NGFragment::Overflow was moved to NGBoxFragment, since neither text nor
linebox has layout overflow. All types of fragments have ink overflow,
but this is not included in this CL.

Some of NGFragmentBuilders could be shared as super classes. This is not
in this CL but to be discussed further.

This patch moves one step towards computing baseline position from
NGFragment, but its implementation is not included in this CL.

BUG=636993

Review-Url: https://codereview.chromium.org/2764753007
Cr-Original-Commit-Position: refs/heads/master@{#459446}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9872e64c2d620fe4c01d7fd1e3776954b162fc9b