Avoid inline capacity of WordMeasurements
This patch disables the inline capacity of WordMeasurements but to
reserve the heap once per line, in order to avoid stack overflow in
LayoutBlockFlow::layoutRunsAndFloatsInRange() when stack is critical,
such as when V8 initiates the layout.

Currently, LayoutBlockFlow::layoutRunsAndFloatsInRange() is the only
place we instantiate WordMeasurements.

This change is expected to save stack space by ~2k.

BUG=704536

Review-Url: https://codereview.chromium.org/2765353004
Cr-Original-Commit-Position: refs/heads/master@{#459485}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 704ca0cfcc046d199d31e6502639bc94b89fddb0