Make resource lookup more uniform in SVGResources::buildResources
Replace the pattern:

if (!ensureResources(resources).setFoo(
        getLayoutSVGResourceById<LayoutSVGResourceBar>(...)))
  treeScopeResources.addPendingResource(...);

with a new pattern using a new attachToResource helper which folds in
the last step (addPendingResource) as well. This makes for less
callsites to adjust when modifying resource lookup.
This also makes the return value of the setFoo(...) methods unnecessary,
so make them return void.

BUG=454767

Review-Url: https://codereview.chromium.org/2772773005
Cr-Original-Commit-Position: refs/heads/master@{#459533}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d7b6fa21c13aebade124c00d08dfc4e19dcf48d3