Moved display property to be generated in ComputedStyleBase.
Move the noninherited property display and its enum, EDisplay,
to be generated in ComputedStyleBase. Also moved getter / setter /
initial value methods to be generated in ComputedStyleBase.

Note that this renames m_effectiveDisplay to m_display to match
the name of the getter/setter. This shouldn't affect anything
since the member is private and only used within the generated
class.

BUG=628043

Review-Url: https://codereview.chromium.org/2672693002
Cr-Original-Commit-Position: refs/heads/master@{#459683}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4efed9fa047ba47f6f4849a597c49a741fafbd18