Stop having separate timeout timers for the preflight and the actual request
DocumentThreadableLoader stops the timeout timer and restarts it with fresh
timeout when getting a preflight response and sending the actual request. This
fixes it so that the timeout is applied to the combined request, not to each of
the preflight and the actual request.

BUG=699985
R=tyoshino@chromium.org

Review-Url: https://codereview.chromium.org/2755483006
Cr-Original-Commit-Position: refs/heads/master@{#459728}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a299a05414fc647524bf1da64fa2b058e5915d69