v8binding: Adds a use counter for named access on window.
* NamedAccessOnWindow_ChildBrowsingContext = 1871
Total count for named access on window to a child browsing context.

* NamedAccessOnWindow_ChildBrowsingContext_CrossOriginNameMismatch = 1872
Count for named access on window to a cross-origin child browsing
context with a different browsing context name.

We'd like to see the ratio, how often the named access would
be denied if we fix crbug.com/538562 .

Once we fix iframe.name so that it updates the browsing context
name, the ratio should be low enough.

BUG=703619

Review-Url: https://codereview.chromium.org/2778723002
Cr-Original-Commit-Position: refs/heads/master@{#459841}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f9df31e030a462ef2421b188c7aa5129c7b6044c