[MD Extensions] Add developerPrivate support for retrying unpacked loads
If an unpacked load fails, we allow the user to "retry", loading the
extension from the same directory (the flow being that the user sees the
error, fixes it, and then hits retry rather than having to re-select the
extension path). Add support for retrying a given load to the
developerPrivate API.

This is a little complicated because, for security reasons, we don't
allow loading an unpacked extension with a path specified by JavaScript.
Instead, associate each load with a unique id, and, on failure, pass
that id to the caller. Then, the caller can specify this retryId in a
subsequent call to loadUnpacked, in which case the file picker will be
skipped and the path associated with that id will be used.

Add tests for the same.

BUG=529395
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/2771243002
Cr-Original-Commit-Position: refs/heads/master@{#459884}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3809d702fc9fb1158849acc4dc5ffe226e4bf662