Rename LayoutObject's paintInvalidationParent() to parentCrossingFrames()
A LayoutObject's paint invalidation parent no longer has special-cases
for multicolumn and is now just the LayoutObject parent with a special-case
that it crosses frames. This patch renames the function which also
improves the readability of non-invalidation callers.

While we're in the neighborhood, slowPaintInvalidationParentForTesting has
been removed because it is not slow and not used.

Review-Url: https://codereview.chromium.org/2783573002
Cr-Original-Commit-Position: refs/heads/master@{#460145}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93efceba4cc433d79d7daeb3fd8d03e1833b2115