Initialize referenceDecodedAudioBuffer correctly.
During the conversion to testharness, we forgot to initialize
referenceDecodedAudioBuffer to the decoded audio file.  This causes
an error to be signaled as a harness error but allowed the tests to
pass still.

Also remove from TestExpectations because the test harness error is
fixed.

BUG=705621, 705125
TEST=decode-audio-data-basic.html

Review-Url: https://codereview.chromium.org/2782443003
Cr-Original-Commit-Position: refs/heads/master@{#460148}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fa85c2e913c2293062171ce3ef31c44216c8b444